backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.32
Flask-Cors==5.0.0
Authlib==1.3.1
requests==2.32.3
typing_extensions==4.12.2
